backend/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
joblib>=1.1.0
fastapi
uvicorn
python-multipart

firebase-admin
